scripts/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
pyyaml==6.0.1
pytest==8.0.0
